+++ virtual_art_museum/data_aquisition/met_museum.py
@@ -1,47 +1,6 @@
 import requests
 import random
-
-class Piece:
-    def __init__(self, id, response):
-        self.id = id
-        self.data = {}
-        self.process_response(response)
-
-    def process_response(self, response):
-        keys = ['title',
-                'artistDisplayName',
-                'image_url', 
-                'department',
-                'objectBeginDate', 
-                'objectEndDate',
-                'medium', 
-                'artistDisplayBio', 
-                'artistNationality']
-
-        for key in keys:
-            if key == 'image_url':
-                self.data[key] = self.check_image_fields(response)
-            else:
-                self.data[key] = response[key]
-
-    def check_image_fields(self, response):
-        if response['primaryImage'] == '':
-            if response['primaryImageSmall'] != '':
-                return response['primaryImageSmall']
-            if response['additionalImages'] != []:
-                return response['additionalImages'][0]
-        else:
-            return response['primaryImage']
-
-    def __str__(self):
-        str =  f"Title: {self.data['title']} \n"
-        str += f"Artist: {self.data['artistDisplayName']} \n"
-        str += f"Department: {self.data['department']} \n"
-        str += f"Image: {self.data['image_url']} \n"
-        str += f"Date: {self.data['objectBeginDate']} - {self.data['objectEndDate']} \n"
-        str += f"Medium: {self.data['medium']} \n"
-        str += f"Display Bio: {self.data['artistDisplayBio']} \n"
-        return str
+from piece import Piece
 
 class MetMuseum:
     def __init__(self, url):
@@ -59,15 +18,11 @@ class MetMuseum:
         if self.check_status(response):
             ids = response.json()['objectIDs']
             pieces = [self.create_piece(id) for id in ids]
-            for piece in pieces:
-                print(piece)
         else:
             return []
 
     def get_N_random_objects(self, N):
         pieces = [self.create_piece(id) for id in random.sample(self.all_ids, N) if self.create_piece(id) is not None]
-        for piece in pieces:
-            print(piece)
         return pieces
 
     def get_all_object_ids(self): #probably wont use this -> it will take forever to get all the objects
